Refactor code in main.rb and create new file called library.rb
